Fix buggs
