+++ .travis.yml
@@ -14,7 +14,7 @@ before_install:
   - octave --eval "pkg install -forge -verbose statistics"
 install:
   - pip install pytest-cov coveralls
-  - pip install -r requirements.txt --no-deps
+  - pip install -r requirements.txt
   - pip install -e .
 script:
   - py.test --cov-config .coveragerc --cov-report term-missing --cov=kgp tests/
